CONTINUE:: ConstructAnAPC_ <- ResolveIDConfOfAPC <- ResolveBidirectionalAxis

+++ core/headers/NeuromorphicTimeSpace/FabricTableConstructors/FabricConstructor.h
@@ -13,7 +13,7 @@ namespace PredictedAdaptedEncoding
         packed64_t* SlabBasePtr_{nullptr};
 
         size_t SlabCellCount_{UNSIGNED_ZERO};
-        uint64_t PerAPCRuntimeCellCount_{UNSIGNED_ZERO};
+        uint16_t PerAPCRuntimeCellCount_{UNSIGNED_ZERO};
         uint64_t CountOfAPC_{UNSIGNED_ZERO};
         uint8_t SlabId_{UNSIGNED_ZERO};
